move sched binary
