+++ Makefile
@@ -29,7 +29,7 @@ vet:
 	go vet ./...
 
 install:
-	go install ./binaries/...
+	go install ./...
 
 # Cleans go.mod and go.sum of unused dependencies
 tidy:
