+++ scheduler/api/thrift/errors.go
@@ -1,4 +1,4 @@
-package api
+package thrift
 
 import (
 	"time"
