+++ scheduler/api/thrift/get_scheduler_status.go
@@ -1,4 +1,4 @@
-package api
+package thrift
 
 import (
 	"github.com/twitter/scoot/sched/scheduler"
