+++ scheduler/api/thrift/get_scheduler_status_test.go
@@ -1,4 +1,4 @@
-package api
+package thrift
 
 import (
 	"testing"
