+++ scheduler/api/thrift/get_status_property_test.go
@@ -1,6 +1,6 @@
 // +build property_test
 
-package api
+package thrift
 
 import (
 	"testing"
