+++ scheduler/api/thrift/kill_job_test.go
@@ -1,4 +1,4 @@
-package api
+package thrift
 
 import (
 	"fmt"
