+++ scheduler/api/thrift/reinstate_worker.go
@@ -1,4 +1,4 @@
-package api
+package thrift
 
 import (
 	"fmt"
