+++ scheduler/api/thrift/run_job_test.go
@@ -1,4 +1,4 @@
-package api
+package thrift
 
 import (
 	"errors"
