+++ scheduler/api/thrift/set_scheduler_status_test
@@ -1,4 +1,4 @@
-package api
+package thrift
 
 import (
 	"strings"
